pyproject.toml
[build-system]
requires = ["setuptools>=64,<81", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "posetail"
version = "0.8.0"
dependencies = [
    "aniposelib>=0.7.0",
    "easydict==1.13",
    "einops==0.7.0",
    "frozendict==2.4.6",
    "opencv-python==4.9.0.80",
    "pytorch-memlab==0.3.0",
    "toml==0.10.2",
    "wandb==0.19.5",
    "wget==3.2",
    "lightning>=2.5.6,<3",
    "rerun-sdk[notebook]<=0.22.1",
    "numpy==1.26",
    "imgaug", "imagecorruptions",
    "setuptools<81",
    "hiera-transformer", "timm",
    "sam-2 @ git+https://github.com/facebookresearch/sam2.git"
    ]
requires-python = ">=3.8"
description = "a model for tracking 2d or 3d animal pose through time"
readme = "README.md"
license = {text = "BSD-3-Clause"}
keywords = ["3d point tracking", "2d point tracking", "multi-view tracking", "pose estimation"]
classifiers = [
    'Intended Audience :: Science/Research',
    'Topic :: Scientific/Engineering :: Image Recognition',
    'License :: OSI Approved :: BSD License',
    'Programming Language :: Python :: 3'
]

# [project.optional-dependencies]
# gpu = [
#   "torch==2.0.1+cu118"
# ]

[tool.pixi.system-requirements]
libc = { family = "glibc", version = "2.34" }

[project.urls]
Repository = "https://github.com/AllenNeuralDynamics/posetail"


[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
posetail = { path = ".", editable = true }
torch = { version = "<3", index="https://download.pytorch.org/whl/cu128"}
torchvision = {version ="*", index="https://download.pytorch.org/whl/cu128"}
torchaudio = {version ="*", index="https://download.pytorch.org/whl/cu128"}

[tool.pixi.environments]
default = { solve-group = "default" }
# gpu = { features = ["gpu"], solve-group = "default" }
dev = {features = ["dev"], solve-group = "default"}

[tool.pixi.tasks]

[tool.pixi.dependencies]
python = "3.10.*"
cuda-toolkit = "12.8.*"
numpy = "==1.26"
setuptools = "<81"
pip = "*"

[tool.pixi.feature.dev.dependencies]
matplotlib = "*"
ipython = "*"
notebook = "*"
ipympl = "*"
av = "*"